miura fold working, added some nice debug actions
Added scrubbing of a playbook and I'm also able to highlight a set of polygons of time now.
So I can enter a polygon index and if it is available it will be highlighted if not - well then not.

With this I can check if polygon is in place at a specified folding step.